+++ src/main/resources/db/migration/20180306114934415__renamed_facility_type_unique_code_constraint.sql
@@ -1,2 +1,4 @@
-ALTER TABLE ONLY facility_types
-    RENAME CONSTRAINT uk_nfppl8ui0vgjoxenm5v2727wo TO unq_facility_type_code;
+ALTER TABLE facility_types
+    DROP CONSTRAINT uk_nfppl8ui0vgjoxenm5v2727wo;
+
+CREATE UNIQUE INDEX unq_facility_type_code ON facility_types (LOWER(code));